+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "Project Timer" extension will be documented in this file.
 
+# [Unreleased]
+### Fixed
+- Cannot activate successfully on windows due to parent path parse error.
+
 ## [0.3.1] - 2026-2-23
 ### Fixed
 - Optimized the refresh logic of status bar and menu after `project-timer.deleteAllStorage`, `project-timer.importData` and `project-timer.renameProject`.